requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
supabase==2.8.1
python-dotenv==1.0.1
pydantic[email]==2.10.3
bcrypt==4.2.1
python-multipart==0.0.12
openai==1.12.0